+++ src/dynamics/hole_filling.jl
@@ -1,25 +1,21 @@
-abstract type AbstractHoleFilling end
+abstract type AbstractHoleFilling <: AbstractModelComponent end
 
-export NoHoleFilling
-struct NoHoleFilling <: AbstractHoleFilling end
-NoHoleFilling(SG::SpectralGrid) = NoHoleFilling()
-initialize!(::NoHoleFilling, ::PrimitiveWet) = nothing
-hole_filling!(::AbstractGrid,::NoHoleFilling,::PrimitiveWet) = nothing
+# no hole filling
+hole_filling!(::AbstractField, ::Nothing, ::AbstractModel) = nothing
 
 export ClipNegatives
 struct ClipNegatives <: AbstractHoleFilling end
 ClipNegatives(SG::SpectralGrid) = ClipNegatives()
-initialize!(::ClipNegatives, ::PrimitiveWet) = nothing
 
 # function barrier
 function hole_filling!(
     q::AbstractField,
     H::ClipNegatives,
-    model::PrimitiveWet
+    model::AbstractModel,
 )
     hole_filling!(q, H)
 end
 
-function hole_filling!(q::AbstractField,::ClipNegatives)
+function hole_filling!(q::AbstractField, ::ClipNegatives)
     @. q = max(q, 0)
 end