+++ src/physics/longwave_radiation.jl
@@ -1,11 +1,8 @@
 abstract type AbstractRadiation <: AbstractParameterization end
 abstract type AbstractLongwave <: AbstractRadiation end
 
-export NoLongwave
-struct NoLongwave <: AbstractLongwave end
-NoLongwave(SG::SpectralGrid) = NoLongwave()
-initialize!(::NoLongwave, ::PrimitiveEquation) = nothing
-longwave_radiation!(::ColumnVariables, ::NoLongwave, ::PrimitiveEquation) = nothing
+# no longwave radiation
+longwave_radiation!(::ColumnVariables, ::Nothing, ::PrimitiveEquation) = nothing
 
 # function barrier for all AbstractLongwave
 function longwave_radiation!(column::ColumnVariables, model::PrimitiveEquation)